requirements.txt
absl-py==2.0.0
alembic==1.12.1
apturl==0.5.2
astunparse==1.6.3
Babel==2.8.0
bcrypt==3.2.0
blinker==1.6.3
Brlapi==0.8.3
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
cliapp==1.20180812.1
click==8.1.7
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
dlib==19.24.2
duplicity==0.8.21
face-recognition==1.3.0
face-recognition-models==0.3.0
fasteners==0.14.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.3
flatbuffers==23.5.26
fpdf==1.7.2
future==0.18.2
gast==0.4.0
gevent==22.10.2
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.59.0
h5py==3.10.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jax==0.4.19
jeepney==0.7.1
Jinja2==3.1.2
keras==2.12.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.3
ml-dtypes==0.3.1
mock==4.0.3
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
panda==0.3.1
pandas==2.1.1
paramiko==2.9.3
pbr==5.8.0
pexpect==4.8.0
Pillow==9.5.0
protobuf==4.24.4
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
QtPy==2.0.0
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.3
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
SQLAlchemy==2.0.23
systemd-python==234
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
ttystatus==0.38
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
Werkzeug==2.2.3
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.event==5.0
zope.interface==6.1
